- Working Event system - Removed SPDLOG for a custom simpler logging system - Working Time system - Working Layer system

+++ OrbitalEngine/src/Platform/OpenGL/OpenGLVertexBuffer.h
@@ -12,9 +12,9 @@ namespace OrbitalEngine
 
 		virtual void bind() const override;
 		virtual void unbind() const override;
-		virtual void allocateMemory(const void* data, unsigned int size) const override;
-		virtual void setAttribute(unsigned int index, unsigned int count, unsigned int type,
-			unsigned int normalized, unsigned int size, const void* pointer) const override;
+		virtual void allocateMemory(const void* data, size_t size) const override;
+		virtual void setAttribute(size_t index, size_t count, unsigned int type,
+			unsigned int normalized, size_t size, const void* pointer) const override;
 
 	private:
 		friend VertexBuffer;
